Split files into independent modules
