+++ lib/read-bitmap.js
@@ -1,0 +1,6 @@
+var fs = require('fs');
+
+module.exports = function readBitmap (file) {
+  var data = fs.readFileSync(file);
+  return data;
+};
